Split studyFactory to controller, resolver, types
